(cms): Cleanup Carousel
