--- apps/cms/src/components/Carousel/EmblaCarouselArrowButtons.tsx
@@ -1,99 +1,0 @@
-import React, {
-  PropsWithChildren,
-  useCallback,
-  useEffect,
-  useState,
-} from 'react';
-import { EmblaCarouselType } from 'embla-carousel';
-
-type UsePrevNextButtonsType = {
-  prevBtnDisabled: boolean;
-  nextBtnDisabled: boolean;
-  onPrevButtonClick: () => void;
-  onNextButtonClick: () => void;
-};
-
-export const usePrevNextButtons = (
-  emblaApi: EmblaCarouselType | undefined
-): UsePrevNextButtonsType => {
-  const [prevBtnDisabled, setPrevBtnDisabled] = useState(true);
-  const [nextBtnDisabled, setNextBtnDisabled] = useState(true);
-
-  const onPrevButtonClick = useCallback(() => {
-    if (!emblaApi) return;
-    emblaApi.scrollPrev();
-  }, [emblaApi]);
-
-  const onNextButtonClick = useCallback(() => {
-    if (!emblaApi) return;
-    emblaApi.scrollNext();
-  }, [emblaApi]);
-
-  const onSelect = useCallback((emblaApi: EmblaCarouselType) => {
-    setPrevBtnDisabled(!emblaApi.canScrollPrev());
-    setNextBtnDisabled(!emblaApi.canScrollNext());
-  }, []);
-
-  useEffect(() => {
-    if (!emblaApi) return;
-
-    onSelect(emblaApi);
-    emblaApi.on('reInit', onSelect);
-    emblaApi.on('select', onSelect);
-  }, [emblaApi, onSelect]);
-
-  return {
-    prevBtnDisabled,
-    nextBtnDisabled,
-    onPrevButtonClick,
-    onNextButtonClick,
-  };
-};
-
-type PropType = PropsWithChildren<
-  React.DetailedHTMLProps<
-    React.ButtonHTMLAttributes<HTMLButtonElement>,
-    HTMLButtonElement
-  >
->;
-
-export const PrevButton: React.FC<PropType> = (props) => {
-  const { children, ...restProps } = props;
-
-  return (
-    <button
-      className='embla__button embla__button--prev'
-      type='button'
-      {...restProps}
-    >
-      <svg className='embla__button__svg' viewBox='0 0 23 39'>
-        <path
-          fill='currentColor'
-          transform='translate(-855, -230)'
-          d='M857.005,231.479L858.5,230l18.124,18-18.127,18-1.49-1.48L873.638,248Z'
-        />
-      </svg>
-      {children}
-    </button>
-  );
-};
-
-export const NextButton: React.FC<PropType> = (props) => {
-  const { children, ...restProps } = props;
-
-  return (
-    <button
-      className='embla__button embla__button--next'
-      type='button'
-      {...restProps}
-    >
-      <svg className='embla__button__svg' viewBox='0 0 23 39'>
-        <path
-          transform='translate(-855, -230)'
-          d='M857.005,231.479L858.5,230l18.124,18-18.127,18-1.49-1.48L873.638,248Z'
-        />
-      </svg>
-      {children}
-    </button>
-  );
-};
