+++ apps/cms/src/pages/editor/components/EmblaCarousel/EmblaCarousel.css
@@ -1,12 +1,12 @@
-.embla__viewport {
+.editor_embla_viewport {
   overflow: hidden;
 }
-.embla__container {
+.editor_embla_container {
   backface-visibility: hidden;
   display: flex;
   touch-action: pan-y;
 }
-.embla__slide {
+.editor_embla_slide {
   display: flex;
   justify-content: center;
   align-items: center;
@@ -15,26 +15,26 @@
 }
 
 @media (max-width: 768px) {
-  .embla__slide {
+  .editor_embla_slide {
     flex: 0 0 50% !important;
     padding-left: 0.5rem !important;
   }
-  .embla__container {
+  .editor_embla_container {
     margin-left: -0.5rem !important;
   }
 }
 
 @media (max-width: 480px) {
-  .embla__slide {
+  .editor_embla_slide {
     flex: 0 0 100% !important;
     padding-left: 0 !important;
   }
-  .embla__container {
+  .editor_embla_container {
     margin-left: 0 !important;
   }
 }
 
-.embla__button {
+.editor_embla_button {
   position: absolute;
   -webkit-tap-highlight-color: rgba(49, 49, 49, 0.5);
   -webkit-appearance: none;
@@ -57,24 +57,24 @@
   transition: opacity 0.25s 0.5s;
 }
 
-.embla__button--prev {
+.editor_embla_button--prev {
   transform: scaleX(-1) scaleY(1);
   top: calc(50% - 1.2rem);
   left: 0.5rem;
 }
 
-.embla__button--next {
+.editor_embla_button--next {
   transform: scaleX(1) scaleY(1);
   top: calc(50% - 1.2rem);
   right: 0.5rem;
 }
 
-.embla__button:disabled {
+.editor_embla_button:disabled {
   pointer-events: none;
   opacity: 0;
 }
 
-.embla__button__svg {
+.editor_embla_button__svg {
   filter: drop-shadow(0 1px 0 rgba(255, 255, 255, 0.5));
   width: 100%;
   height: 100%;
